+++ src/domain/flow-system/src/entities/flow/flow_outcome.rs
@@ -9,10 +9,11 @@
 
 use kamu_core::PullResult;
 use kamu_task_system as ts;
+use opendatafabric::Multihash;
 
 /////////////////////////////////////////////////////////////////////////////////////////
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+#[derive(Debug, Clone, PartialEq, Eq)]
 pub enum FlowOutcome {
     /// Flow succeeded
     Success(FlowResult),
@@ -24,38 +25,41 @@ pub enum FlowOutcome {
     Aborted,
 }
 
+impl FlowOutcome {
+    pub fn try_result_as_ref(&self) -> Option<&FlowResult> {
+        match self {
+            Self::Success(flow_result) => Some(flow_result),
+            _ => None,
+        }
+    }
+}
+
 /////////////////////////////////////////////////////////////////////////////////////////
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+#[derive(Debug, Clone, PartialEq, Eq)]
 pub enum FlowResult {
     Empty,
     DatasetUpdate(FlowResultDatasetUpdate),
 }
 
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
+#[derive(Debug, Clone, PartialEq, Eq)]
 pub struct FlowResultDatasetUpdate {
-    pub num_blocks: u64,
-    pub num_records: u64,
+    pub old_head: Option<Multihash>,
+    pub new_head: Multihash,
 }
 
-impl From<&ts::TaskResult> for FlowResult {
-    fn from(value: &ts::TaskResult) -> Self {
+impl From<ts::TaskResult> for FlowResult {
+    fn from(value: ts::TaskResult) -> Self {
         match value {
             ts::TaskResult::Empty => Self::Empty,
-            ts::TaskResult::PullResult(pull_result) => match *pull_result {
-                PullResult::UpToDate => Self::DatasetUpdate(FlowResultDatasetUpdate {
-                    num_blocks: 0,
-                    num_records: 0,
-                }),
-                PullResult::Updated {
-                    num_blocks,
-                    num_records,
-                    ..
-                } => Self::DatasetUpdate(FlowResultDatasetUpdate {
-                    num_blocks,
-                    num_records,
-                }),
-            },
+            ts::TaskResult::UpdateDatasetResult(task_update_result) => {
+                match task_update_result.pull_result {
+                    PullResult::UpToDate => Self::Empty,
+                    PullResult::Updated { old_head, new_head } => {
+                        Self::DatasetUpdate(FlowResultDatasetUpdate { old_head, new_head })
+                    }
+                }
+            }
         }
     }
 }