+++ src/infra/core/src/pull_service_impl.rs
@@ -642,8 +642,6 @@ impl PullService for PullServiceImpl {
             Ok(res) => Ok(PullResult::Updated {
                 old_head: Some(res.old_head),
                 new_head: res.new_head,
-                num_blocks: 1,
-                num_records: 0,
             }),
             Err(
                 WriteWatermarkError::EmptyCommit(_)