+++ src/infra/core/tests/tests/test_sync_service_impl.rs
@@ -165,7 +165,7 @@ async fn do_test_sync(
             old_head: None,
             new_head,
             num_blocks: 2,
-            num_records: _,
+            ..
         }) if new_head == b1
     );
 
@@ -175,7 +175,7 @@ async fn do_test_sync(
             old_head: None,
             new_head,
             num_blocks: 2,
-            num_records: _,
+            ..
         }) if new_head == b1
     );
 
@@ -208,7 +208,7 @@ async fn do_test_sync(
             old_head,
             new_head,
             num_blocks: 2,
-            num_records: _,
+            ..
         }) if old_head.as_ref() == Some(&b1) && new_head == b3
     );
 
@@ -218,7 +218,7 @@ async fn do_test_sync(
             old_head,
             new_head,
             num_blocks: 2,
-            num_records: _,
+            ..
         }) if old_head.as_ref() == Some(&b1) && new_head == b3
     );
 
@@ -267,7 +267,7 @@ async fn do_test_sync(
             old_head,
             new_head,
             num_blocks: 1,
-            num_records: _,
+            ..
         }) if old_head == Some(b3.clone()) && new_head == exta_head
     );
 
@@ -296,7 +296,7 @@ async fn do_test_sync(
             old_head,
             new_head,
             num_blocks: 4, // full resynchronization: seed, b1, b2, b3
-            num_records: _,
+            ..
         }) if old_head == Some(exta_head.clone()) && new_head == b3
     );
 
@@ -324,7 +324,7 @@ async fn do_test_sync(
             old_head,
             new_head,
             num_blocks: 4, // full resynchronization: seed, b1, b2, b3
-            num_records: _,
+            ..
         }) if old_head == Some(exta_head.clone()) && new_head == b3
     );
 
@@ -357,7 +357,7 @@ async fn do_test_sync(
             old_head,
             new_head,
             num_blocks: 2,
-            num_records: _,
+            ..
         }) if old_head.as_ref() == Some(&b3) && new_head == b5
     );