+++ src/infra/flow-system-inmem/src/services/flow/active_configs_state.rs
@@ -20,7 +20,7 @@ use crate::dataset_flow_key::*;
 pub(crate) struct ActiveConfigsState {
     dataset_schedules: HashMap<FlowKeyDataset, Schedule>,
     system_schedules: HashMap<SystemFlowType, Schedule>,
-    dataset_start_conditions: HashMap<FlowKeyDataset, StartConditionConfiguration>,
+    dataset_batching_rules: HashMap<FlowKeyDataset, BatchingRule>,
 }
 
 impl ActiveConfigsState {
@@ -34,8 +34,8 @@ impl ActiveConfigsState {
             FlowConfigurationRule::Schedule(schedule) => {
                 self.dataset_schedules.insert(key, schedule);
             }
-            FlowConfigurationRule::StartCondition(condition) => {
-                self.dataset_start_conditions.insert(key, condition);
+            FlowConfigurationRule::BatchingRule(batching) => {
+                self.dataset_batching_rules.insert(key, batching);
             }
         }
     }
@@ -63,7 +63,7 @@ impl ActiveConfigsState {
 
     fn drop_dataset_flow_config(&mut self, flow_key: BorrowedFlowKeyDataset) {
         self.dataset_schedules.remove(flow_key.as_trait());
-        self.dataset_start_conditions.remove(flow_key.as_trait());
+        self.dataset_batching_rules.remove(flow_key.as_trait());
     }
 
     pub fn try_get_flow_schedule(&self, flow_key: &FlowKey) -> Option<Schedule> {
@@ -79,14 +79,14 @@ impl ActiveConfigsState {
         }
     }
 
-    pub fn try_get_dataset_start_condition(
+    pub fn try_get_dataset_batching_rule(
         &self,
         dataset_id: &DatasetID,
         flow_type: DatasetFlowType,
-    ) -> Option<StartConditionConfiguration> {
-        self.dataset_start_conditions
+    ) -> Option<BatchingRule> {
+        self.dataset_batching_rules
             .get(BorrowedFlowKeyDataset::new(dataset_id, flow_type).as_trait())
-            .cloned()
+            .copied()
     }
 }