+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+### Added
+- Flow system now fully supports batching conditions for derived datasets:
+   - not launching excessive flows unless minimal number of input records is accumulated
+   - not waiting on the batching condition over a limit of 24h, if at least something accumulated
 ### Fixed
 - `kamu login` no longer requires workspace in `--user` scope (#525)
 - Sync will correctly select smart ODF protocol when pushing/pulling via repository alias (#521)